code refactoring and performance improvements

+++ ConsoleRayTracer/Terminal/ITerminal.cs
@@ -2,8 +2,8 @@
 
 interface ITerminal
 {
-    short Width { get; }
-    short Height { get; }
+    int Width { get; }
+    int Height { get; }
 
     void SetPixel(int x, int y, float color);
     void Draw();
